Repository restructured with corrected imports
