+++ funcs/vaultdb.py
